+++ tests/test_converters.py
@@ -1,5 +1,9 @@
 import spdx_license_list
 from vecorel_cli.converters import Converters
+from vecorel_cli.vecorel.schemas import VecorelSchema
+from vecorel_cli.vecorel.util import load_file
+
+from fiboa_cli.fiboa.version import get_fiboa_uri
 
 
 def test_converters(capsys):
@@ -32,3 +36,23 @@ def test_valid_license():
             assert converter.license.upper() in spdx_license_list.LICENSES, (
                 f"Converter {_id} has invalid license {converter.license}"
             )
+        assert getattr(converter, "license") is None or isinstance(converter.license, str)
+        assert getattr(converter, "provider") is None or isinstance(converter.provider, str)
+
+
+def test_overriden_base_properties():
+    """
+    You should not define a different schema for a property if it is defined in the base schema.
+    """
+    c = Converters()
+    for _id in Converters().list_ids():
+        converter = c.load(_id)
+        schemas = converter.missing_schemas
+        converter_properties = schemas and schemas.get("properties") or {}
+        schema = VecorelSchema(load_file(get_fiboa_uri()))
+
+        for property, s in schema["properties"].items():
+            if property in converter_properties:
+                assert s == converter_properties[property], (
+                    "Converter {converter} overrides schema for base property {property}"
+                )