+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Fix to vecorel: converter.license and provider should be string
 - Check for license validity, either SPDX string or custom with url
+- Avoid base property schema override
 
 ## [v0.20.3] - 2025-09-13